requirements.txt
hydra_core==1.3.2
h5py
lightning_utilities==0.14.3
markdown==3.8.2
mpmath==1.3.0
multidict==6.6.0
numpy==2.2.2
omegaconf==2.3.0
pytorch_lightning==2.5.1
PyYAML==6.0.2
scikit_learn==1.5.2
scipy==1.15.1
setproctitle==1.3.4
sklearn==0.0
sympy==1.13.1
tensorboard==2.19.0
tensorboard_data_server==0.7.2
tensorboardX==2.6.2.2
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.14.0